Converted App.css to App.module.css
- Converted App.css to App.module.css
- Import global stylings in App.module.css to components along with their component css module
- CocktailForm styling migrated to its own css module from App.css
