+++ src/components/App/App.module.css
@@ -39,37 +39,13 @@
   }
 }
 
-.background-img {
+.backgroundImage {
   background-size: cover;
   background-repeat: no-repeat;
   position: fixed;
   top: 0;
   left: 0;
   width: 100vw;
-  z-index: -20;
-  height: 100vh;
-  width: 100vw;
-}
-
-#profile-img {
-  background-size: cover;
-  background-repeat: no-repeat;
-  position: fixed;
-  top: 0;
-  left: 0;
-  /* width: 100vw; */
-  z-index: -20;
-  height: 100vh;
-  width: 100vw;
-}
-
-#cocktails-img {
-  background-size: cover;
-  background-repeat: no-repeat;
-  position: fixed;
-  top: 0;
-  left: 0;
-  /* width: 100vw; */
   z-index: -20;
   height: 100vh;
   width: 100vw;
@@ -88,6 +64,14 @@
   z-index: -20;
   min-width: 300px;
   max-height: 2000px;
+}
+
+.formContainer {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
 }
 
 .home-text {
@@ -109,36 +93,6 @@
 
 .home-text:hover{
 background:#a6efed;
-}
-
-.favorite-nav {
-  position: absolute;
-  text-decoration: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: bolder;
-  color: #706897;
-  width: 100px;
-  height: 40px;
-  left: 76%;
-  top: 10px;
-  background: #E5E5E5;
-  border-radius: 24px;
-  cursor: pointer;
-}
-
-.favorite:hover{
-  background: #a6dcef;
-}
-
-.home-page {
-  color: #542e71;
-  padding: 37rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
 }
 
 label {
@@ -228,7 +182,7 @@ input.cocktail-box {
 }
 
 
-.view-more-btn, .view-more-btn:link, .view-more-btn:visited {
+.viewMoreButton, .viewMoreButton:link, .viewMoreButton:visited {
   text-decoration: none;
   padding: 13px 0;
   border: none;
@@ -243,11 +197,9 @@ input.cocktail-box {
   cursor: pointer;
 }
 
-.view-more-btn:hover {
+.viewMoreButton:hover {
   background: rgb(88, 188, 242);
-  /* border: none; */
   color: #fefefe;
-  /* border-radius: 20px; */
 }
 
 
@@ -294,50 +246,6 @@ input.cocktail-box {
 
 .image-container img:hover .image-container{
   height: 200px;
-}
-
-.cocktail-detail {
-  display: grid;
-  grid-template: 1fr/ 27vw 20vw 1fr;
-  margin-top: 80px;
-  grid-column-gap: 50px;
-}
-
-.cocktail-detail-1 {
-  margin-left: 50px;
-  grid-area: 1/1/-1/2;
-}
-
-.cocktail-detail-1 img {
-  width: 70%;
-  /* max-height: 270px; */
-  border-radius: 15px;
-}
-
-.cocktail-detail-2 {
-  grid-area: 1/2/-1/3;
-}
-
-.cocktail-detail-1-category {
-  font-weight: 600;
-}
-
-.cocktail-detail-2 img {
-  width: 70%;
-  border-radius: 15px;
-}
-
-.follow-btn {
-  color: #706897;
-  background-color: #E7D9EA;
-  width: 130px;
-  border-radius: 20px;
-  padding: 7px;
-  font-weight: 600;
-  margin: 5px;
-  margin-top: 10px;
-  opacity: 0.8;
-  cursor: pointer;
 }
 
 .cocktail-detail-3 {
