+++ src/components/Home/Home.module.css
@@ -1,10 +1,9 @@
-.img {
-    background-size: cover;
-        background-repeat: no-repeat;
-        position: fixed;
-        top: 0;
-        left: 0px;
-        z-index: -10;
-        height: 100vh;
-        width: 100vw;
+.container {
+    color: #542e71;
+    padding: 37rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
 }
