+++ src/features/categories/CategoryCard/CategoryCard.jsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom'
-import Styles from './CategoryCard.module.css'
+import styles from './CategoryCard.module.css'
 
 const CategoryCard = ({ category }) => {
 	const { id, name, definition, popular_drinks } = category
@@ -7,22 +7,22 @@ const CategoryCard = ({ category }) => {
 		popular_drinks[Math.floor(Math.random() * popular_drinks.length)]
 
 	return (
-		<div className={Styles.card}>
-			<div className={Styles.image}>
+		<div className={styles.card}>
+			<div className={styles.image}>
 				<img
 					src={categoryCocktail.image}
 					alt={categoryCocktail.name}
 					loading="lazy"
 				/>
-				<div className={Styles.overlay}> {categoryCocktail.name}</div>
+				<div className={styles.overlay}> {categoryCocktail.name}</div>
 			</div>
-			<div className={Styles.info}>
+			<div className={styles.info}>
 				<Link to={`/categories/${id}`}>
 					<h3>{name}</h3>
 				</Link>
 				<p>{definition}</p>
 			</div>
-			<Link to={`/categories/${id}`} className={Styles.viewMore}>
+			<Link to={`/categories/${id}`} className={styles.viewMore}>
 				View More
 			</Link>
 		</div>
