+++ src/features/categories/CategoryCard/CategoryCard.module.css
@@ -1,7 +1,7 @@
 .card {
     display: flex;
     flex-direction: column;
-    justify-content: start;
+    justify-content: flex-start;
     align-items: center;
     min-height: 500px;
     width: auto;
