+++ src/features/cocktails/CocktailDetail/CocktailDetail.module.css
@@ -1,0 +1,231 @@
+.container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: flex-start;
+    flex-wrap: wrap;
+    margin-top: 4rem;
+    gap: 2rem;
+    padding: 8rem 0 2rem 4rem;
+}
+
+.cocktail {
+    flex: 2;
+    display: flex;
+    flex-direction: row;
+    gap: 2rem
+}
+
+.image {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+    gap: 1rem;
+}
+
+.image img {
+    width: 70%;
+    border-radius: 0.5rem;
+}
+
+.about {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+}
+
+.container h3 {
+    font-size: 2rem;
+    font-weight: bolder;
+    color: #706897;
+    margin: 0;
+}
+
+.category {
+    margin-top: 1rem;
+}
+.category span {
+    font-weight: 600;
+}
+
+.bartender {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center
+}
+
+.avatar {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+    gap: 1rem;
+}
+
+.avatar img {
+    width: 50%;
+    border-radius: 0.5rem;
+}
+
+.info {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+    gap: 0;
+}
+
+.followButton {
+    color: #706897;
+    background-color: #E7D9EA;
+    width: 130px;
+    border-radius: 20px;
+    padding: 7px;
+    font-weight: 600;
+    margin: 5px;
+    margin-top: 10px;
+    opacity: 0.8;
+    cursor: pointer;
+}
+
+@media (max-width: 500px) {
+    .container {
+        flex-direction: column;
+        padding: 0 2rem 2rem;
+        gap: 4rem;
+    }
+    .container h3 {
+        font-size: 1.5rem;
+    }
+    .container p,
+    .container span {
+        font-size: 14px;
+    }
+    .cocktail {
+        flex-direction: column;
+        gap: 1rem;
+    }
+    .about h3 {
+        display: none;
+    }
+    .image  {
+        flex-direction: column-reverse;
+    }
+}
+
+@media (min-width: 500px) and (max-width: 800px) {
+    .container {
+        padding: 4rem 0 0 2.5rem;
+        gap: 1rem
+    }
+    .cocktail {
+        gap: 3rem;
+    }
+    .image img {
+        width: 100%;
+    }
+    .about h3 {
+        display: none;
+    }
+    .about p:first-of-type{
+        margin-top: 0;
+    }
+    .avatar img {
+        width: 70%;
+    }
+    .bartender {
+        gap: 1rem;
+    }
+}
+
+@media (min-width: 800px) and (max-width: 1100px) {
+    .container {
+        padding: 6rem 0 0 2.5rem;
+        gap: 1rem
+    }
+    .cocktail {
+        gap: 3rem;
+    }
+    .image img {
+        width: 100%;
+    }
+    .avatar img {
+        width: 70%;
+    }
+    .bartender {
+        gap: 1rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 600px) and (max-width: 800px) {
+    .container {
+        padding: 0 0 0 1.5rem;
+        gap: 0
+    }
+    .container h3 {
+        font-size: 1.25rem;
+    }
+    .container p,
+    .container span {
+        font-size: 12px;
+    }
+    .cocktail {
+        gap: 2rem;
+    }
+    .image img {
+        width: 80%;
+    }
+    .avatar img {
+        width: 60%;
+    }
+    .bartender {
+        gap: 1rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 800px) and (max-width: 1000px) {
+    .container {
+        padding: 0 0 0 3rem;
+        gap: 0
+    }
+    .container h3 {
+        font-size: 1.75rem;
+    }
+    .cocktail {
+        gap: 3rem;
+    }
+    .image img {
+        width: 90%;
+    }
+    .avatar img {
+        width: 60%;
+    }
+    .bartender {
+        gap: 1rem;
+    }
+}
+
+@media (orientation: landscape) and (min-width: 1000px) and (max-width: 1200px) {
+    .container {
+        padding: 4rem 0 0 4rem;
+        gap: 0
+    }
+    .cocktail {
+        gap: 3rem;
+    }
+    .image img {
+        width: 90%;
+    }
+    .avatar img {
+        width: 60%;
+    }
+    .bartender {
+        gap: 1rem;
+    }
+}
