+++ src/features/cocktails/CocktailForm/CocktailForm.module.css
@@ -1,0 +1,98 @@
+.form {
+    opacity: 85%;
+    background: white;
+    padding: 2rem 5rem 1rem;
+    border-radius: 20px;
+    box-shadow: 10px 10px 10px 1px rgba(0, 0, 0, 0.2);
+}
+
+.form h1 {
+    width: 100%;
+    color: #706897;
+    border-bottom: 2px solid #706897;
+    line-height: 30px;
+    width: 100%;
+    font-weight: bolder;
+    text-align: center;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding-bottom: 0.5rem;
+}
+
+.form label {
+    display: block;
+    margin: 1.5rem 0 0.5rem
+}
+
+.form textarea,
+.form input {
+    width: 100%;
+    border: 1px solid lightgrey;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 20px;
+    padding: 8px;
+    letter-spacing: 1px;
+    font-weight: 600;
+    font-size: 11pt;
+    background: transparent;
+    transition: 1s;
+    outline: none;
+}
+
+.form button {
+    width: 50%;
+    justify-content: center;
+    align-items: center;
+    background-color: #F3C397;
+    opacity: 0.7;
+    color: #706897;
+    margin-top: 5px;
+    border-radius: 20px;
+    border: none;
+    font-size: 15pt;
+    font-weight: 700;
+    transition: 1s;
+    outline: none;
+    cursor: pointer;
+    margin-top: 20px;
+    margin-bottom: 20px;
+    padding: 7px;
+}
+
+.form select {
+    width: 100%;
+    border: 1px solid lightgrey;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 20px;
+    padding: 2px;
+    padding-left: 10px;
+    letter-spacing: 1px;
+    font-weight: 600;
+    margin: 5px;
+    font-size: 11pt;
+    background: transparent;
+    transition: 1s;
+    outline: none;
+    appearance: none;
+}
+
+.form select:focus {
+    border-color: #706897;
+    outline: none;
+}
+/* Remove the blue outline on select elements in Firefox */
+.form select:-moz-focusring {
+    color: transparent;
+    text-shadow: 0 0 0 #000;
+}
+
+.form select option {
+    padding: 0.5em;
+    color: #706897;
+    background-color: white;
+}
